move map tests to own file
